trying to fix compression

+++ optimization/shared/fed_avg_schedule_ou_compression.py
@@ -419,7 +419,7 @@ def build_fed_avg_process(
   # server_state_type = initialize_computation.type_signature.result
   # model_weights_type = server_state_type.model
   predicted_delta_type = type_conversions.type_from_tensors(_get_weights(dummy_model).trainable) 
-  round_num_type = tf.int32
+  round_num_type = tf.float32
   threshold_type = tf.float32
   num_participants_type = tf.int32
   norm_mean_type = tf.float32
